+++ src/ducks/alertSnackbar.js
@@ -1,3 +1,4 @@
+// constants
 const SHOW_ALERT_SNACKBAR = 'rhoades/alertSnackbar/SHOW_ALERT_SNACKBAR';
 const HIDE_ALERT_SNACKBAR = 'rhoades/alertSnackbar/HIDE_ALERT_SNACKBAR';
 
@@ -14,6 +15,7 @@ const initialState = {
   options: defaultOptions,
 };
 
+// reducer
 export default function reducer(state = initialState, action) {
   switch (action.type) {
     case SHOW_ALERT_SNACKBAR:
@@ -33,6 +35,7 @@ export default function reducer(state = initialState, action) {
   }
 }
 
+// actions
 export const showAlertSnackbar = (messageParams) => ({
   type: SHOW_ALERT_SNACKBAR,
   payload: { ...messageParams },